requirements.txt
colorlog==5.0.1
cachetools==4.2.2
certifi==2020.12.5 
chardet==4.0.0 
google-auth==1.30.0 
idna==2.10 
kubernetes==17.17.0 
oauthlib==3.1.0 
pyasn1==0.4.8 
pyasn1-modules==0.2.8 
python-dateutil==2.8.1 
pyyaml==5.4.1 
requests==2.25.1 
requests-oauthlib==1.3.0 
rsa==4.7.2 
setuptools==56.2.0 
six==1.16.0 
urllib3==1.26.4 
websocket-client==1.0.0
numpy==1.19.5
pandas==1.1.5
pytz==2021.1
argparse==1.4.0
deepdiff==5.5.0
ordered-set==4.0.2
